+++ lib-ui-core/src/main/java/com/pax/us/pay/ui/core/api/IRespStatus.java
@@ -3,5 +3,5 @@ package com.pax.us.pay.ui.core.api;
 public interface IRespStatus {
     void onAccepted();
 
-    void onDeclined(int code, String message);
+    void onDeclined(long code, String message);
 }
